# docs/design/battery-storage-discovery/openapi.yaml
...
"openapi": "3.0.3"
"info":
  "title": "Pandora Battery Storage & Traceability — MVP API v1"
  "version": "1.1.0"
  "description": "Обновлённый проект по решению руководителя: шкафы, полки, ячейки отдельными таблицами не ведутся. Адреса — текстовые source_location/destination_location в immutable журнале и current_location в текущей проекции. Одна АКБ на точный канонический адрес; формат адреса: три положительных числа шкаф.полка.ячейка без ведущих нулей. У строк нет FK на существующие места, нет управления пустыми/отключёнными местами. RETURN только взявшим сотрудником. Детальный клиент ТСД/offline вне этапа. Device code — необязательная серверная метка, без справочника устройств. POST/PATCH изменения требуют Idempotency-Key, lookup read-only без ключа. Четыре domain таблицы плюс техническая idempotency_requests. Authentication provider и источник сотрудников ещё не подтверждены. Этот YAML встроен в новый MVP backend. Локальный режим использует явно заданный Bearer test token."
"servers":
  -
    "url": "/api/v1"
    "description": "Тот же origin, что и Swagger UI"
"security":
  -
    "ApiBearer": []
...
"paths":
  "/employees":
    "get":
      "operationId": "listEmployees"
      "tags":
        - "Employees"
      "summary": "Список сотрудников"
      "description": "Список сотрудников"
      "responses":
        "200":
          "description": "Успешное чтение."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/EmployeePage"
              "example":
                "items":
                  -
                    "id": "00000000-0000-4000-8000-000000000137"
                    "is_active": true
                    "created_at": "2026-09-30T09:14:00Z"
                    "updated_at": "2026-09-30T09:14:00Z"
                    "disabled_at": null
                    "display_name": "Иван Петров"
                    "personnel_number": "004281"
                "next_cursor": null
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
      "parameters":
        -
          "$ref": "#/components/parameters/Limit"
        -
          "$ref": "#/components/parameters/Cursor"
        -
          "name": "is_active"
          "in": "query"
          "schema":
            "type": "boolean"
    "post":
      "operationId": "createEmployee"
      "tags":
        - "Employees"
      "summary": "Создать сотрудника"
      "description": "Не выполняется автоматически после неизвестного скана. Требуется право ведения справочника; корпоративный source employees требует Q05; MVP поддерживает тестовый локальный реестр."
      "responses":
        "201":
          "description": "Успешный результат. Replay возвращает первоначальный результат; GET нужен для свежего состояния."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Employee"
              "example":
                "id": "00000000-0000-4000-8000-000000000137"
                "is_active": true
                "created_at": "2026-09-30T09:14:00Z"
                "updated_at": "2026-09-30T09:14:00Z"
                "disabled_at": null
                "display_name": "Иван Петров"
                "personnel_number": "004281"
          "headers":
            "Idempotency-Replayed":
              "description": "true при возврате сохранённого результата"
              "schema":
                "type": "boolean"
            "Location":
              "description": "URL созданного ресурса или /api/v1/operations/{operation_id} для движения"
              "schema":
                "type": "string"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
      "parameters":
        -
          "$ref": "#/components/parameters/IdempotencyKey"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/CreateEmployee"
            "example":
              "display_name": "Иван Петров"
              "personnel_number": "004281"
  "/employees/{employee_id}":
    "get":
      "operationId": "getEmployee"
      "tags":
        - "Employees"
      "summary": "Сотрудник по ID"
      "description": "Сотрудник по ID"
      "responses":
        "200":
          "description": "Успешное чтение."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Employee"
              "example":
                "id": "00000000-0000-4000-8000-000000000137"
                "is_active": true
                "created_at": "2026-09-30T09:14:00Z"
                "updated_at": "2026-09-30T09:14:00Z"
                "disabled_at": null
                "display_name": "Иван Петров"
                "personnel_number": "004281"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
      "parameters":
        -
          "$ref": "#/components/parameters/employee_id"
    "patch":
      "operationId": "patchEmployee"
      "tags":
        - "Employees"
      "summary": "Изменить имя или активность"
      "description": "Нельзя изменить ID. Предлагается запрещать отключение с выданными АКБ: 409 EMPLOYEE_HAS_CUSTODY."
      "responses":
        "200":
          "description": "Успешный результат. Replay возвращает первоначальный результат; GET нужен для свежего состояния."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Employee"
              "example":
                "id": "00000000-0000-4000-8000-000000000137"
                "is_active": true
                "created_at": "2026-09-30T09:14:00Z"
                "updated_at": "2026-09-30T09:14:00Z"
                "disabled_at": null
                "display_name": "Иван Петров"
                "personnel_number": "004281"
          "headers":
            "Idempotency-Replayed":
              "description": "true при возврате сохранённого результата"
              "schema":
                "type": "boolean"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
      "parameters":
        -
          "$ref": "#/components/parameters/employee_id"
        -
          "$ref": "#/components/parameters/IdempotencyKey"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/PatchEmployee"
            "example":
              "is_active": false
  "/credential-resolutions":
    "post":
      "operationId": "resolveCredential"
      "tags":
        - "Employees"
      "summary": "Найти человека по активному credential"
      "description": "Read-only POST ради отсутствия credential в URL. Не создаёт сотрудника, не требует Idempotency-Key. Credential не аутентифицирует API-клиента."
      "responses":
        "200":
          "description": "Успешное чтение."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/CredentialResolution"
              "example":
                "employee":
                  "id": "00000000-0000-4000-8000-000000000137"
                  "is_active": true
                  "created_at": "2026-09-30T09:14:00Z"
                  "updated_at": "2026-09-30T09:14:00Z"
                  "disabled_at": null
                  "display_name": "Иван Петров"
                  "personnel_number": "004281"
                "credential_id": "00000000-0000-4000-8000-000000000138"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/ResolveCredential"
            "example":
              "credential_value": "0004097281923"
  "/employees/{employee_id}/credentials":
    "get":
      "operationId": "listCredentials"
      "tags":
        - "Employees"
      "summary": "Карты сотрудника, включая историю"
      "description": "Только право ведения справочников; значения credential не публикуются для произвольного читателя."
      "responses":
        "200":
          "description": "Успешное чтение."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/CredentialPage"
              "example":
                "items":
                  -
                    "id": "00000000-0000-4000-8000-000000000138"
                    "is_active": true
                    "created_at": "2026-09-30T09:14:00Z"
                    "updated_at": "2026-09-30T09:14:00Z"
                    "disabled_at": null
                    "employee_id": "00000000-0000-4000-8000-000000000137"
                    "value": "0004097281923"
                "next_cursor": null
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
      "parameters":
        -
          "$ref": "#/components/parameters/employee_id"
        -
          "$ref": "#/components/parameters/Limit"
        -
          "$ref": "#/components/parameters/Cursor"
    "post":
      "operationId": "createCredential"
      "tags":
        - "Employees"
      "summary": "Назначить или заменить карту"
      "description": "Несколько active credential поддерживаются. replaces_credential_id отключает конкретную старую в той же транзакции. 409 ACTIVE_CREDENTIAL_EXISTS/CREDENTIAL_REUSE_UNCONFIRMED."
      "responses":
        "201":
          "description": "Успешный результат. Replay возвращает первоначальный результат; GET нужен для свежего состояния."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Credential"
              "example":
                "id": "00000000-0000-4000-8000-000000000138"
                "is_active": true
                "created_at": "2026-09-30T09:14:00Z"
                "updated_at": "2026-09-30T09:14:00Z"
                "disabled_at": null
                "employee_id": "00000000-0000-4000-8000-000000000137"
                "value": "0004097281923"
          "headers":
            "Idempotency-Replayed":
              "description": "true при возврате сохранённого результата"
              "schema":
                "type": "boolean"
            "Location":
              "description": "URL созданного ресурса или /api/v1/operations/{operation_id} для движения"
              "schema":
                "type": "string"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
      "parameters":
        -
          "$ref": "#/components/parameters/employee_id"
        -
          "$ref": "#/components/parameters/IdempotencyKey"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/CreateCredential"
            "example":
              "value": "0004097281924"
              "replaces_credential_id": "00000000-0000-4000-8000-000000000138"
  "/employees/{employee_id}/credentials/{credential_id}":
    "patch":
      "operationId": "disableCredential"
      "tags":
        - "Employees"
      "summary": "Деактивировать credential"
      "description": "Меняется только активность. Value и employee_id неизменяемы. Отключение уже отключённой карты возвращает 200 без изменения исходного disabled_at."
      "responses":
        "200":
          "description": "Успешный результат. Replay возвращает первоначальный результат; GET нужен для свежего состояния."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Credential"
              "example":
                "id": "00000000-0000-4000-8000-000000000138"
                "is_active": false
                "created_at": "2026-09-30T09:14:00Z"
                "updated_at": "2026-09-30T09:14:00Z"
                "disabled_at": "2026-09-30T09:14:00Z"
                "employee_id": "00000000-0000-4000-8000-000000000137"
                "value": "0004097281923"
          "headers":
            "Idempotency-Replayed":
              "description": "true при возврате сохранённого результата"
              "schema":
                "type": "boolean"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
      "parameters":
        -
          "$ref": "#/components/parameters/employee_id"
        -
          "$ref": "#/components/parameters/credential_id"
        -
          "$ref": "#/components/parameters/IdempotencyKey"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/DisableCredential"
            "example":
              "is_active": false
  "/batteries":
    "get":
      "operationId": "listBatteries"
      "tags":
        - "Batteries"
      "summary": "Поиск и текущие состояния"
      "description": "Поиск и текущие состояния"
      "responses":
        "200":
          "description": "Успешное чтение."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/BatteryPage"
              "example":
                "items":
                  -
                    "id": "00000000-0000-4000-8000-000000000412"
                    "inventory_code": "B-00412"
                    "serial_number": null
                    "status": "STORED"
                    "current_location": "1.1.1"
                    "current_holder_employee_id": null
                    "version": 1
                    "created_at": "2026-09-30T09:14:00Z"
                    "updated_at": "2026-09-30T09:14:00Z"
                "next_cursor": null
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
      "parameters":
        -
          "$ref": "#/components/parameters/Limit"
        -
          "$ref": "#/components/parameters/Cursor"
        -
          "name": "inventory_code"
          "in": "query"
          "schema":
            "type": "string"
            "minLength": 1
          "description": "Точное сравнение, 0 или 1 результат; физический формат маркировки неизвестен"
        -
          "name": "status"
          "in": "query"
          "schema":
            "type": "string"
            "enum":
              - "STORED"
              - "ISSUED"
        -
          "name": "holder_employee_id"
          "in": "query"
          "schema":
            "type": "string"
            "format": "uuid"
        -
          "name": "location"
          "in": "query"
          "schema":
            "type": "string"
            "pattern": '^[1-9][0-9]*\.[1-9][0-9]*\.[1-9][0-9]*$'
            "minLength": 1
          "description": "Точный текстовый адрес. В batteries — current_location, в operations — source OR destination. Не подтверждает существование места."
    "post":
      "operationId": "storeNewBattery"
      "tags":
        - "Commands"
      "summary": "Зарегистрировать и разместить АКБ"
      "description": "Атомарный реестр + STORE version=1. destination_location непустая, без крайних пробелов, должна быть свободна по UNIQUE. Формат адреса: три положительных числа через точку без ведущих нулей. Существование физического места не проверяется справочником. 409 INVENTORY_CODE_EXISTS/LOCATION_OCCUPIED."
      "responses":
        "201":
          "description": "Успешный результат. Replay возвращает первоначальный результат; GET нужен для свежего состояния."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/CommandResult"
              "example":
                "battery":
                  "id": "00000000-0000-4000-8000-000000000412"
                  "inventory_code": "B-00412"
                  "serial_number": null
                  "status": "STORED"
                  "current_location": "1.1.1"
                  "current_holder_employee_id": null
                  "version": 1
                  "created_at": "2026-09-30T09:14:00Z"
                  "updated_at": "2026-09-30T09:14:00Z"
                "operation":
                  "id": "00000000-0000-4000-8000-000000000901"
                  "battery_id": "00000000-0000-4000-8000-000000000412"
                  "battery_version": 1
                  "type": "STORE"
                  "actor_employee_id": "00000000-0000-4000-8000-000000000137"
                  "credential_id": "00000000-0000-4000-8000-000000000138"
                  "source_status": null
                  "destination_status": "STORED"
                  "source_location": null
                  "destination_location": "1.1.1"
                  "source_holder_employee_id": null
                  "destination_holder_employee_id": null
                  "device_code": "TSD-example"
                  "occurred_at": "2026-09-30T09:14:00Z"
          "headers":
            "Idempotency-Replayed":
              "description": "true при возврате сохранённого результата"
              "schema":
                "type": "boolean"
            "Location":
              "description": "URL созданного ресурса или /api/v1/operations/{operation_id} для движения"
              "schema":
                "type": "string"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
        "409":
          "$ref": "#/components/responses/Error409"
        "413":
          "$ref": "#/components/responses/Error413"
        "415":
          "$ref": "#/components/responses/Error415"
      "parameters":
        -
          "$ref": "#/components/parameters/IdempotencyKey"
      "requestBody":
        "required": true
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/StoreBattery"
            "example":
              "inventory_code": "B-00412"
              "actor_credential_value": "0004097281923"
              "destination_location": "1.1.1"
  "/batteries/{battery_id}":
    "get":
      "operationId": "getBattery"
      "tags":
        - "Batteries"
      "summary": "Текущее состояние АКБ"
      "description": "Текущее состояние АКБ"
      "responses":
        "200":
          "description": "Успешное чтение."
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/Battery"
              "example":
                "id": "00000000-0000-4000-8000-000000000412"
                "inventory_code": "B-00412"
                "serial_number": null
                "status": "STORED"
                "current_location": "1.1.1"
                "current_holder_employee_id": null
                "version": 1
                "created_at": "2026-09-30T09:14:00Z"
                "updated_at": "2026-09-30T09:14:00Z"
        "400":
          "$ref": "#/components/responses/Error400"
        "401":
          "$ref": "#/components/responses/Error401"
        "403":
          "$ref": "#/components/responses/Error403"
        "404":
          "$ref": "#/components/responses/Error404"
        "422":
          "$ref": "#/components/responses/Error422"
        "500":
          "$ref": "#/components/responses/Error500"
        "503":
          "$ref": "#/components/responses/Error503"
      "parameters":
        -
          "$ref": "#/components/parameters/battery_id"
...
"components":
  "securitySchemes":
    "ApiBearer":
      "type": "http"
      "scheme": "bearer"
      "description": "Предложенный доверенный API principal/сеанс, без утверждения JWT/OIDC или конкретного issuer. Employee credential не заменяет token. Provider и права требуют Q19–20/Q25."
...
  "responses":
    "Error400":
      "description": "Некорректный запрос"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
          "example":
            "error":
              "code": "INVALID_REQUEST"
              "message": "Некорректный запрос"
              "details": {}
              "request_id": "trace-example"
    "Error401":
      "description": "Требуется аутентификация"
      "content":
        "application/json":
          "schema":
            "$ref": "#/components/schemas/Error"
          "example":
            "error":
              "code": "UNAUTHENTICATED"
              "message": "Требуется аутентификация"
              "details": {}
              "request_id": "trace-example"
...
  "schemas":
...
    "EmployeePage":
      "type": "object"
      "additionalProperties": false
      "properties":
        "items":
          "type": "array"
          "items":
            "$ref": "#/components/schemas/Employee"
        "next_cursor":
          "type": "string"
          "nullable": true
      "required":
        - "items"
        - "next_cursor"
...